requirements.txt
aiohttp>=3.8.0
python-dotenv>=0.19.0
requests>=2.26.0
asyncio>=3.4.3
logging>=0.5.1.2
typing>=3.7.4.3
datetime>=4.3
json>=2.0.9
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0 